context board, context modal and editabText added

+++ client/src/graphql/mutations/columnMutations.js
@@ -23,7 +23,7 @@ export const UPDATE_COLUMN = gql`
     $id: ID!
     $title: String
     $boardId: ID
-    $order: Int
+    $order: Float
   ) {
     updateColumn(
       _id: $id
